refactor: book cards
